requirements.txt
anyio==3.6.2
attrs==22.1.0
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
fastapi==0.85.1
filelock==3.8.0
greenlet==1.1.3.post0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
packaging==21.3
pbr==5.11.0
platformdirs==2.5.4
pluggy==1.0.0
psycopg2-binary==2.9.4
py==1.11.0
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.40
starlette==0.20.4
stevedore==4.1.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
virtualenv==20.14.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
